feat: Emit valid RUM view events
refs: RUM-12321

This PR updates `RumViewScope` to produce actual RUM View events in lieu
of the `{"placeholder-for":"view"}` values it was previously emitting.
View events are populated with a basic subset of properties to describe
the view.

A prior PR added exhaustive tests to `RumViewEvent` itself: those tests
validate that the JSON serialization logic for `RumViewEvent` is fully
sound, and that the resulting JSON payloads are valid according to the
schema in `rum-events-format`.

This PR updates the existing RUM tests as well:

- In the API tests, we validate that RUM produces the expected set of
  events in response to API calls

- In the implementation-layer tests for `RumViewScope`, we validate that
  `RumViewScope` produces the expected set of events in response to
  `RumCommand` values being processed

I've tweaked the `validate-event-types` script a bit as well: we now use
`DATADOG_RUM_EVENT_LITERAL` as the cue that we should validate a JSON
literal as a RUM event, and we also allow a tiny bit of string template
substitution when describing the expected event payload. Specifically,
we can now write `{"session":{"id":"${__NONZERO_UUID__}"}}`, and:

1. The tests will pass so long as `session.id` is a valid non-nil UUID
2. `validate-event-types` will preprocess the value and substitute a
   random UUID value before performing schema validation