chore: In logger API, rename custom-attribute functions for consistency
This PR is a very small refactor that renames a small set of public API
functions: namely, the functions that allow users to manage custom
attributes for log events.

There are two separate naming changes here: one is about consistency
with the existing mobile SDKs, while the other is about consistency
between the C and C++ APIs.

 ### Consistency with mobile SDKs: Set/Delete vs. Add/Remove

I initially used `SetAttribute()` and `DeleteAttribute()` in the logging
API, as it mirrored the JSON object semantics implemented in
`Attribute::Object`, and "Set" felt marginally clearer at conveying that
multiple calls with the same name will replace existing values.

However, the iOS and Android SDKs name these functions `AddAttribute()`
and `RemoveAttribute()`, consistently, across multiple features. There's
no compelling reason to be different in the C++ SDK.

 ### Consistency between C and C++: attribute_add vs. add_attribute

It's a fairly common idiom in C APIs that when you have a series of
functions that operate on some type `mylib_foo_t`, you name those
functions `mylib_foo_<operation>()`. That approach makes sense for
dd-sdk-cpp: where we have `datadog::Logger::Info()` in the C++ API, the
C API equivalent is `dd_logger_info()`.

However, in some places I extended that naming convention a bit too far,
and over time it started to feel wonky in places. I initially had:

- `dd_logger_attribute_add` _(conceptually: invoking the `add` operation
on an `attribute` belonging to a `dd_logger`)_

...rather than:

- `dd_logger_add_attribute` _(conceptually: invoking the `add_attribute`
operation on `dd_logger`)_

To put it another way, which of these options feels better to you?

1. `datadog::Rum::StartView` and `dd_rum_start_view()`
2. `datadog::Rum::StartView` and `dd_rum_view_start()`

If the C function operated on a type called `dd_rum_view_t`, and the
equivalent C++ function was a member of `RumView`, then I'd go with
option 2... but as it stands, I much prefer option 1. I think it's much
clearer if the name of the actual operation is the same in both places.

...and that's a long-winded explanation of why this PR changes the
logging API's `attribute_set`/`attribute_delete` to
`add_attribute`/`remove_attribute`, consistent with a similar change
made in the RUM API. Thanks for coming to my TED Talk.