fix: Replace vector-based ScopeArray with fixed-sized RumViewArray
refs: RUM-12946

This PR removes the generic, `std::vector`-based `ScopeArray` container
implementation in favor of a new container called `RumViewArray`, which
wraps a `std::array<std::optional<RumViewScope>, 8>` and enforces a hard
limit on the number of concurrent view scopes that can exist in a
session.

 ### Why did this need to change?

As it stands, the SDK will crash if:

- More than 8 views are started concurrently
- One or more views are stopped while one or more other,
  more-recently-created views are allowed to persist

The bug here is that we're maintaining persistent references to parent
scopes (e.g. an **Action** keeps a reference to its parent **View**),
but View scopes are stored in a `std::vector`. When the vector is
reallocated, or when elements are erased and rearranged, those
references are invalidated.

See [RUM-12946](https://datadoghq.atlassian.net/browse/RUM-12946) for
more details.

 ### Why is a generic ScopeArray unnecessary?

We only use `ScopeArray<T>` in one place, as `ScopeArray<RumViewScope>`:
there's no need for arrays of arbitrarily-typed scopes beyond this
single usage in `RumSessionScope`.

As a refresher, here are the key parent/child relationships between RUM
view scopes:

- **Application** has 0 or 1 **Sessions**; but may retain a reference to
  the previous session _(two member vars is fine; no container needed)_

- **Session** has 0..n **Views** _(requires an array)_

- **View** has 0 or 1 **Actions** _(one member var is fine; no container
  needed)_

- **View** has 0..n **Resources**, indexed by key _(requires a map)_

This means we only need two container types for RUM scopes:

- An array of `RumViewScope` for use by `RumSessionScope`
- An associative array of `RumResourceScope` for use by `RumViewScope`

In both cases, we want fast insertion and removal, and we don't need to
preserve insertion order.

 ### Why is std::vector not a good fit?

As noted above, we maintain persistent references to scopes, so we can't
store scope values directly in a `std::vector`. If we want stable
references, we have to make a choice on this tradeoff:

1. Perform a heap allocation for every new scope, and store pointers
2. Place a hard limit on the number of scopes that can be maintained

This PR represents the latter approach.

Re: memory overhead: `sizeof(RumViewScope)` is currently about ~296
bytes, so `sizeof(RumViewArray)` is ~2.5 kB. This cost is baked into the
initial heap allocation made when initializing `Rum`, such that we never
need to call `malloc` simply because a new view was started.

 ### Why is a hard limit of 8 permissible?

The active session will generally only have a single view scope, as only
one view can be active at any given time. Scopes for inactive views only
hang around if there are pending resources within that view.

So an application would hit the view scope limit if it:

- Started a view
- Started one or more resources within that view
- Transitioned to another view and repeated ~8 times without ever
  recording the end of those pending resources

While this can happen, it seems unlikely enough that some data loss
would be acceptable. If this happens, the oldest view that's still
waiting for resource completion will be dropped (without sending a final
view event; similar to what happens if the session expires while
resources are pending) and a warning will be logged.