feat: Fully define RUM resource commands
refs: RUM-12202

> This PR is an incremental step toward implementing RUM resources:
> resource changes ripple out to a lot of existing code/tests, and
> resources and errors have a bit of overlap, so I'm attempting to open
> incremental, bite-sized PRs to keep reviews manageable.

This PR adds the full set of resource-related properties to the existing
`StartResource` and `StopResource` commands. There are a few common sets
of fields that can be used to describe a resource (HTTP request details,
HTTP response details, and error details), so these are defined as
structs within the implementation layer.