feat: Add per-command profiling support to repl
refs: RUM-13346

This PR adds the following repl commands:

- `nop [label]`: Does nothing; prints label as output if provided
- `start-profile [cpu|memory]`: Begins profiling each subsequent command
- `stop-profile`: Stops profiling

 ### Standard I/O for portability

All profiling data is printed directly to stdout. This means we can
adapt our automated profiling workflow to any platform, including
support for profiling on remote devices, so long as we can:

1. Compile the SDK and repl for that platform
2. Write a script file to the remote device (if necessary)
3. Remotely launch the repl process on that process and either:
    - Point it to our script file, or
    - Pipe commands to it directly over stdin
4. Capture stdout from that process

 ### Rationale

The repl profiling workflow is intended for running **automated,
repeatable benchmarks** so that we can resolve key metrics that describe
the performance characteristics of any given revision of the SDK,
allowing us to monitor those metrics and detect performance regressions.

For interactive profiling that simulates a real-world app, with minimal
profiler overhead and with deep insights into SDK threads and
callstacks: that's a different use case, and it would be better served
using an instrumented profiler like Tracy.

 ### Profiling behavior

In `cpu` mode, the elapsed time to run each command (in nanoseconds)
will be printed immediately following the result of the command:

```
> log hello
< Logger::Log()
<<(duration): 375
```

This figure represents the time taken to execute the API call identified
in the output, plus the minuscule overhead of parsing the input string
and branching to the correct repl command-handler function. It is not
contaminated with any overhead from console I/O.

Running in `memory` mode adds considerable slowdown, but it will also
print detailed information about heap allocations made during the
execution of the command, and by any background threads for 1ms
thereafter:

```
> log hello
< Logger::Log()
<<(duration): 375
<<(alloc-events): 6
<<<(0): op:alloc size:128 tid:3241635350488800441
<<<(1): op:alloc size:4096 tid:5376371447635715523
<<<(2): op:free size:4096 tid:5376371447635715523
<<<(3): op:alloc size:88 tid:5376371447635715523
<<<(4): op:free size:88 tid:5376371447635715523
<<<(5): op:free size:128 tid:5376371447635715523
```

When `start-profile` is run, it prints the main thread's `tid` value
(e.g. `<<<(main-thread-id): 3241635350488800441`), allowing the main
thread to be differentiated from any background threads.

When running interactively, note that this output will include any
allocs/frees that have occurred since the last command. The `nop` and
`sleep` commands are useful for flushing accumulated allocation events
before running the next command.

Allocation tracking is handled via a overloads for the `new` and
`delete` operators, using code that was previously authored for unit
tests.

Currently, after calling `start-profile`, a maximum of 4096 allocation
events (`new` or `delete` calls) may be recorded. If this limit is
reached, `stop-profile` will result in an error, failing the script
invocation rather than allowing truncated data to be used.