feat: Add an app for visualizing repl benchmark data
refs: RUM-13351

This PR adds a "viewer" application to `tools/repl-profile`. This is a
quick-and-dirty application intended to visualize profiling data
collected from repl benchmarks.

[]

Loading up the app will allow you to choose a benchmark, platform,
and build config, along with a revision to view data for (and optionally
another revision to treat as a baseline). If profiling data is available
for the selection revision(s), the app will display:

- A plot of the main-thread CPU time required to execute each command in
  the benchmark (i.e. how long the calling thread was blocked)

- A plot of the number of allocations made in response to each command
  execution

- A chart showing current heap size at the time each command was
  executed

- A listing of all new/delete calls made for a selected command, with
  number of bytes allocated/freed and an ID for the calling thread

This app is built with Streamlit, an open-source immediate-mode UI
framework for quickly building data-sciencey apps that can easily render
Python datasets and charts.

Note that running the main repl-profile script (to generate and store
data) does not require any third-party dependencies. Running the viewer
app requires installing the dependencies listed in `requirements.txt`.