chore: Delete the old benchmark program
refs: RUM-13351

The `dd_native_benchmark` program, built from source in `benchmarks/`
when `DD_BUILD_BENCHMARKS` is enabled, represents an earlier attempt to
create automated benchmarks that could be run as a single C++ program
without external dependencies.

The newer workflow (using `repl` in conjunction with external Python
scripts, which are much faster to develop and which don't pollute the
address space of the process being profiled) seems like a better
approach to lead with.

Since these old benchmarks aren't of much use in their current form, I
think we should delete them to avoid confusion (and make our builds a
bit smaller to boot).